CORE - optimalization to max duration work.

+++ core/src/net/personaltt/simplesolver/SolverSolution.java
@@ -15,7 +15,7 @@ public interface SolverSolution {
      * Returns current cost
      * @return 
      */
-    public int cost();
+    public long cost();
     
     
     /**
